LinkedList implementation of IntoIterator trait
The trait is implemented for the list's reference and the list's mutable
reference. It's not implemented for the list itself, because the list does
not own its items anyway. It only owns mutable references to items. So
it does not make sense to consume the list to iterate over mutable
references.