Include interior mutability for Mutex
This patch creates interior mutability using RefCell. Possibly it will be
replaced with UnsafeCell in the future when more sophisticated detection
of broken critical section will be introduced in the Mutex module.

From the Mutex user perspective anything dereferenced to the type contained
by mutex must be returned from borrow or borrow_mut functions. It should
not make any assumptions what exact type it is.