Refactor Timer traits
Make Timer traits cleaner to use.
This refactoring was triggered by an attempt to create a method to check if
a scheduled hardware task was trigger. But meantime it turned out that there
is no easy reliable way to implement such check with a freerunning timer.

So the result is cleaner trait interface, but no new feature